+++ tests/testthat/test-scatterplotmatrix_app.R
@@ -5,7 +5,8 @@
 
 tns <- tns_factory("not_ebas")
 
-ID <- poc( # nolint
+ID <- poc(
+  # nolint
   INPUT = poc(
     CAT = tns(SPM$ID$PAR, "cat_val"),
     PAR = tns(SPM$ID$PAR, "par_val"),
@@ -19,19 +20,19 @@ ID <- poc( # nolint
   )
 )
 
+skip_if_not_running_shiny_tests()
+
 root_app <- start_app_driver(dv.explorer.parameter::mock_app_scatterplotmatrix())
 on.exit(if ("stop" %in% names(root_app)) root_app$stop())
 
-fail_if_app_not_started <- function() {
-  if (is.null(root_app)) rlang::abort("App could not be started")
+
+if (is.null(root_app)) {
+  rlang::abort("App could not be started")
 }
 
-test_that("data is subset according to selection", {
-  testthat::skip_if_not(run_shiny_tests)
-  fail_if_app_not_started()
-  skip_if_suspect_check()
-
+local({
   app <- shinytest2::AppDriver$new(root_app$get_url())
+  app_args <- dv.explorer.parameter::mock_app_boxplot(dry_run = TRUE)
 
   inputs <- list()
   inputs[[ID$INPUT$CAT]] <- "PARCAT2"
@@ -48,94 +49,61 @@ test_that("data is subset according to selection", {
   app$set_inputs(!!!inputs2)
   app$wait_for_idle()
 
-  testthat::skip_if_not(run_shiny_tests)
-  fail_if_app_not_started()
-  skip_if_suspect_check()
+  exported_test_values <- app$get_value(export = "not_ebas-output_arguments")
 
-  data_subset <- app$get_value(export = tns("data_subset"))
-
-  expect_true(
-    all(inputs2[[ID$INPUT$PAR]] %in% names(data_subset))
+  expected_ds <- spm_subset_data(
+    cat = inputs[[ID$INPUT$CAT]],
+    cat_col = app_args$srv$cat_var,
+    par = inputs2[[ID$INPUT$PAR]],
+    par_col = app_args$srv$par_var,
+    val_col = inputs[[ID$INPUT$VAL]],
+    vis = inputs[[ID$INPUT$VIS]],
+    vis_col = app_args$srv$visit_var,
+    group_vect = stats::setNames(c(inputs[[ID$INPUT$GRP]]), c(CNT$MAIN_GROUP)),
+    bm_ds = shiny::isolate(app_args$srv$bm_dataset()),
+    group_ds = shiny::isolate(app_args$srv$group_dataset()),
+    subj_col = app_args$srv$subjid_var
   )
 
-  expect_equal(
-    attr(data_subset[[CNT$MAIN_GROUP]], "label"),
-    paste0("Label of ", inputs[[ID$INPUT$GRP]])
+  test_that(
+    "scatterplotmatrix chart is included according to selection" |>
+      vdoc[["add_spec"]](c(
+        specs$scatterplot_matrix_module$composition,
+        specs$scatterplot_matrix_module$scatterplot_matrix_chart
+      )),
+    {
+      expected <- list(
+        ds = expected_ds
+      )
+
+      exported <- shiny::isolate(exported_test_values[[SPM$ID$CHART]][["arguments"]]())
+
+      expect_identical(exported, expected)
+
+      plot <- app$get_value(output = ID$OUTPUT$CHART)
+      expect_true("src" %in% names(plot))
+    }
   )
 
-  expect_equal(
-    unique(as.character(data_subset[[CNT$VIS]])),
-    inputs[[ID$INPUT$VIS]]
+  test_that(
+    "bookmark is restored. Clicks are excluded are not bookmarked." |>
+      vdoc[["add_spec"]](c(specs$scatterplot_matrix_module$bookmark)),
+    {
+      bmk_url <- suppressWarnings(app$get_js("window.location.href"))
+
+      bookmark_app <- shinytest2::AppDriver$new(bmk_url)
+      bookmark_app$wait_for_idle()
+      app_input_values <- app$get_values()[["input"]]
+
+      # Explicitely exclude click
+      bmk_input_values <- bookmark_app$get_values()[["input"]]
+      expect_identical(app_input_values, bmk_input_values)
+    }
   )
 })
 
-test_that("scatterplotmatrix chart is included according to selection" |>
-  vdoc[["add_spec"]](c(specs$scatterplot_matrix_module$composition, specs$scatterplot_matrix_module$scatterplot_matrix_chart)), {
-  testthat::skip_if_not(run_shiny_tests)
-  fail_if_app_not_started()
-  skip_if_suspect_check()
-
-  app <- shinytest2::AppDriver$new(root_app$get_url())
-
-  inputs <- list()
-  inputs[[ID$INPUT$CAT]] <- "PARCAT2"
-  inputs[[ID$INPUT$VAL]] <- "VALUE2"
-  inputs[[ID$INPUT$VIS]] <- "VISIT2"
-  inputs[[ID$INPUT$GRP]] <- "CAT3"
-
-  # Set in two steps because a prior selector is required
-  inputs2 <- list()
-  inputs2[[ID$INPUT$PAR]] <- c("PARAM22", "PARAM23")
-
-  app$set_inputs(!!!inputs)
-  app$wait_for_idle()
-  app$set_inputs(!!!inputs2)
-  app$wait_for_idle()
-
-  chart <- app$get_html(paste0("#", ID$OUTPUT$CHART))
-  expect_snapshot(chart, cran = TRUE)
-})
-
-test_that("bookmark is restored. Clicks are excluded are not bookmarked." |>
-  vdoc[["add_spec"]](c(specs$scatterplot_matrix_module$bookmark)), {
-  testthat::skip_if_not(run_shiny_tests)
-  fail_if_app_not_started()
-  skip_if_suspect_check()
-
-  app <- shinytest2::AppDriver$new(root_app$get_url())
-
-  inputs <- list()
-  inputs[[ID$INPUT$CAT]] <- "PARCAT2"
-  inputs[[ID$INPUT$VAL]] <- "VALUE2"
-  inputs[[ID$INPUT$VIS]] <- "VISIT2"
-  inputs[[ID$INPUT$GRP]] <- "CAT3"
-
-  # Set in two steps because a prior selector is required
-  inputs2 <- list()
-  inputs2[[ID$INPUT$PAR]] <- c("PARAM22", "PARAM23")
-
-  app$set_inputs(!!!inputs)
-  app$wait_for_idle()
-  app$set_inputs(!!!inputs2)
-  app$wait_for_idle()
-
-  # URL is automatically updated with the bookmarked URL
-  bmk_url <- app$get_js("window.location.href")
-
-  bookmark_app <- shinytest2::AppDriver$new(bmk_url)
-  bookmark_app$wait_for_idle()
-  app_input_values <- app$get_values()[["input"]]
-
-  # Explicitely exclude click
-  bmk_input_values <- bookmark_app$get_values()[["input"]]
-  expect_identical(app_input_values, bmk_input_values)
-})
 
 test_that("default values are set", {
-  testthat::skip_if_not(run_shiny_tests)
-  fail_if_app_not_started()
-  skip_if_suspect_check()
-
   srv_defaults <- list(
     default_cat = "PARCAT2",
     default_par = c("PARAM22", "PARAM23"),
@@ -163,13 +131,7 @@ test_that("default values are set", {
 })
 
 
-
-
 test_that("default values are set including analysis flag variables", {
-  testthat::skip_if_not(run_shiny_tests)
-  fail_if_app_not_started()
-  skip_if_suspect_check()
-
   srv_defaults <- list(
     default_cat = "PARCAT2",
     default_par = c("PARAM22", "PARAM23"),