+++ tests/testthat/test-wfphm_hmpar.R
@@ -235,9 +235,9 @@ local({
         )
       ),
     {
-      testthat::skip_if_not(run_shiny_tests)
+      skip_if_not_running_shiny_tests()
       fail_if_app_not_started()
-      skip_if_suspect_check()
+      
       expect_r2d3_svg(app, list(list(svg = C$SVG_JS_QUERY, container = C$CONTAINER, n = 2)))
     }
   )
@@ -245,9 +245,9 @@ local({
   test_that(
     "wfphm-hmpar returns margin list",
     {
-      testthat::skip_if_not(run_shiny_tests)
+      skip_if_not_running_shiny_tests()
       fail_if_app_not_started()
-      skip_if_suspect_check()
+      
       x <- app$get_values()
       margin <- shiny::isolate(x[["export"]][[tns("r")]][["margin"]]())
       expect_true(setequal(names(margin), c("top", "bottom", "left", "right")))
@@ -261,9 +261,9 @@ local({
       )
       ),
     {
-      testthat::skip_if_not(run_shiny_tests)
+      skip_if_not_running_shiny_tests()
       fail_if_app_not_started()
-      skip_if_suspect_check()
+      
 
       expect_identical(app$get_values()[["export"]][["not_ebas-chart_args"]][["x_axis"]], "S")
     }