+++ NEWS.md
@@ -1,3 +1,7 @@
+# dv.explorer.parameter 0.1.9-9010
+
+Refactor tests
+
 # dv.explorer.parameter 0.1.9-9009
 
 Remove other menu from scatterplot matrix